pyproject.toml
[project]
name = "facompbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Jean Carlos",email = "jeancc.costa@gmail.com"},
    {name = "Emilly Caroline",email = "07emillycaroline07@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-adk (>=1.20.0,<2.0.0)",
    "litellm (>=1.80.7,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]

[tool.poetry]
packages = [{include = "facompbot", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
